delete functionality for orders and changed orderId to firebaseKey in orders.json
