+++ src/scripts/components/orders.js
@@ -18,7 +18,7 @@ const showOrders = (array) => {
         <div class="card-body">
           <button type="button" class="btn btn-outline-primary">Details</button>
           <button type="button" class="btn btn-outline-info">Edit</button>
-          <button type="button" class="btn btn-outline-danger">Delete</button>
+          <button type="button" id="delete-order--${item.firebaseKey}" class="btn btn-outline-danger">Delete</button>
         </div>
       </div>`;
   });
