--- src/scripts/helpers/data/odersData.js
@@ -1,27 +1,0 @@
-import axios from 'axios';
-import firebaseConfig from '../../../api/apiKeys';
-
-const dbUrl = firebaseConfig.databaseURL;
-
-const getOrders = () => new Promise((resolve, reject) => {
-  axios.get(`${dbUrl}/orders.json`)
-    .then((response) => resolve(Object.values(response.data)))
-    .catch((error) => reject(error));
-});
-// GET SINGLE ORDER
-const getSingleOrder = (firebaseKey) => new Promise((resolve, reject) => {
-  axios.get(`${dbUrl}/orders/${firebaseKey}.json`)
-    .then((response) => resolve(response.data))
-    .catch(reject);
-});
-// DELETE
-const deleteOrder = (firebaseKey) => new Promise((resolve, reject) => {
-  axios.delete(`${dbUrl}/orders/${firebaseKey}.json`)
-    .then(() => {
-      getOrders().then(resolve);
-    })
-    .catch(reject);
-});
-// GET ORDER ITEMS FROM SINGLE ORDER
-
-export { getOrders, getSingleOrder, deleteOrder };
